Refactoring of build scripts
The Core, Common and Functions sub-projects each create their own JAR,
and a meta-build script calls their build scripts in sequence. Since
there is a chain of dependencies (Common depends on Core, Functions
depends on Common), the meta-build script copies the JAR compiled at the
previous step in the `lib` folders of the other sub-projects.