Refactoring of ExplanationQueryable
The interface now requires a RelationNodeFactory instead of a plain NodeFactory. This is an API-breaking change for all classes that directly implement the interface or override the getExplanation() method from an ancestor. All code in this commit is left unchanged with the exception of the reference to the new class name. Modifications that require the use of the RelationNodeFactory are left for a subsequent commit.